Fix exp_month error when Stripe is selected before changing to a different payment processor.  Issue #16342.
The Stripe extension adds custom javascript validation to the contribution
page billing fields. Unfortunately, it doesn't do a good job of preventing
the validation from firing if there are multiple payment processors in use.
Basically, if Stripe is selected and then a different payment processor is
selected, it will still try to perform the javascript validation, even though
Stripe isn't the selected payment processor.

This commit fixes this issue. It's a bit hacky, but instead of binding the
form's submit event and performing the validation that way, we're binding to
a click event on the submit button. In addition, we're only binding to that
button if the form has a class of "stripe-payment-form", which gets removed
every time the payment processor is changed and only added back in Stripe is
selected.  Seems to work fine this way.