The previous attempt at fixing the PayPal/Strip exp_month error failed.
Civi doesn't have a great way of handling custom js validation per payment
processor, so I have to take the nuclear option of completely unbinding the
js submit event handler that Stripe is adding every time the payment
processor is switched on the contribution page.  If it's switched to Stripe,
the code will re-add the submit event handler.  Sigh.