ZUse X-Forwarded-* headers to determine base_url.
We are running this site behind varnish which sets the X-Forwarded-Host header
to what the original HTTP Host header was. We want to use this for the base
URL instead of setting the base URL in the settings.php file because then
we don't need to worry about changing the settings.php file for the staging
or dev sites.