Fixed Stripe.js token error.
I'm still not exactly sure what the problem was. Normally, the Main
donation page (the first one) should have a hidden stripe_token field in
it and that gets stored in the form's session in Civi. Adding that field
gets handled by the buildForm hook in the stripe module. I think it
wasn't adding the field because in this case the form has two payment
processors and it defaults to PayPal (not Stripe), so when the page
loads originally, Stripe isn't set as the payment processor so the
stripe buildForm hook doesn't add the stripe_token field. When you click
on 'Credit Card' (Stripe) for the payment processor, that causes an AJAX
snippet request which triggers that buildForm hook, but it doesn't add
the field because for this AJAX request it is setting type=11.

What I don't understand about this fix: 1. Why it was working sometimes
for us during testing. 2. What that empty($_GET['type']) check was for
and will we break something else by removing it.