Check mailing list boxes by default - Issue #16368
Request came from client to have the CC Events and CC Newsletter subscription
opt-ins to be checked by default in the donation form. Could not find any way
to do so through the GUI, so added a bit of javascript in the theme to check
the boxes on page load.