Log all stripe errors.
Some people are getting errors when trying to donate and we know that
stripeCatchErrors is geting an error because of the information the user
reported, but we don't know what specifically is the problem. This adds
logging on any error in stripeCatchErrors which hopefully will give us some
idea about what's going on.