Allow campaign_id to be blank. Issue #16860
It's not clear to my why, but it looks like the Stripe extension requires
that recurring contributions have a campaign set. This was causing webhooks
from Stripe to error out because we don't have a campaign set on the
contribution page. I don't see why we would need to require Stripe
contributions to use campaigns, so I adjusted it so that it allows NULL
for the campaign_id field.