Add Media Browser Plus to default install profile. Issue #22163.
I also cleaned up the make file a bit as we don't need to define the
directory for every module since we can set a default. The syntax for modules
where we're not specifying a version is inconsistent with how we currently
have things, so I left the directory syntax in place for any modules where
we're not specifying a version to download as I think it's easier to read.