Draft GA eCommerce Tracking Issue #22660
This commit includes a draft version of the Google Analytics Tracking
Code CC requested we add to the Contribution Thank You page.  It
includes debug information, so that Anthony can help troubleshoot the
ORDER_NUMBER field (and possible SOURCE_CODE) which is not populating
with the Contribution ID of the Contribution just processed.