Moving Billing Block Changes Issue #22669
Instead of checking the Billing Same As Shipping and changing the label
with JS, we are instead customizing the Stripe Extension's template.
This allows these elements to load on the page with these changes,
instead of users experiencing a slight delay.  And was the cleanest
option, as the Stripe Extension overwrote the general, CiviCRM custom
template for the Billing Block.