Chaning Zen Gutters Issue #22661
Previously, the gutter between Zen Grid columns was being set on
main_content.  This meant that even when the page was down to one
column, the donate form had a 10px margin on the left and a 30px margin
on the right.  Now, the gutter is set in sidebars.scss and only when Zen
is using multiple columns.