Autopopulating Billing Address fields Issue #22669
Anthony noticed that simply loading the Billing Block with Same as
Shipping checked, does not autopopulate the fields with a user's Shipping
information.  So I have instead set the box as checked in JS and asked
that it trigger the function filling out the Billing fields with a user's Shipping
information.