Remove "every" rather than replacing recurring HTML code. Issue #22663.
JQuery's .html() function should be used sparingly, especially around forms.
It works by replacing the entire DOM node with the HTML specified, which
means that, even if most of the HTML you're replacing it with remains the
same, any JavaScript event bindings on those elements will be gone. In this
case, we're probably fine, but the result could be that if there were an
onclick event binded to anything within the recurring contribution HTML, it
wouldn't work anymore.