Styling Amount options Issue #22665
This commit styles the Contribution Amount options, so that they appear
as green buttons someone can select.  Once selected, it will change
color, and all other options will return to their unselected state. I
also changed the button colors on the page because the previous greens
were conflicting with the heart and pie chart, as well as the yellow
green colors in the header and footer.