Updating ID on recurring label Issue #22667
Eric noticed that the label for the recurring donation option was not
displaying the word "monthly."  This appears to have been caused by the
live site having a different ID for that field than my local copy.  I
have updated the ID in script.js to match the live site.