Default if no other premium selected Issue #23345
Anthony requested that we only set the "No thank you" default when no
other premium is selected on page load.  This helps save someone's
preference in the event they hit an error on submission and the page
automatically reloads.  Previously, it would have lost their selection
and defaulted to "No thank you" again.