Change event to fill Billing fields Issue #23648
I noticed on subsequent testing that the Billing Fields were only
filling the first digit of Zip Code because I was using the browser's
autofill feature and Zip Code was the last field in the Shipping Field
group.  Input appears to update the Billing Fields even when
completing the Shipping Fields with a browser autofill feature.