Change event and check "Same as" Issue #23648
Input was not working as an event to fill Billing fields with Shipping
information.  So I changed the event to "change" on all Shipping fields
and now check whether the box "Billing same as Shipping" is checked
before filling the Billing fields with any changes made to the Shipping
fields.