Turn off disable logic for alternate processors
When you have a PayPal processor with a Stripe processor both enabled on
the page, if the user is using PayPal and clicks on submit, this code
would disable the button and it wouldn't work again without a cache
refresh. We just want the standard behaviour if not using the Stripe
processor.