Fixed problem where strip token wasn't getting set
I was getting the "Stripe.js token was not passed!" error if I wasn't
changing the price from the default when using Stripe. This turned out to
be because the skipPaymentMethod javascript in CiviCRM changed how it
calculates currentTotal and it wasn't working the way the Stripe extension
was expecting it to. It looks like in both 4.5 and 4.6 Civi has been
hiding .payment_options-group so we'll use that to figure out if it was
free so that we don't have to copy the complicated price calculation code.