Set HTTP status code to 500 on error
When there was an error processing the stripe message, the HTTP status
code was getting set to 200. Stripe interpets that as the webhook
succeeding so it won't retry or show any error in its interface. This
changes things so that any exception will get logged and tell Stripe there
was an error.