Ported billing field changes
The newer version of the Strip payment processor doesn't override the
default CiviCRM BillingBlock template. Instead it uses some javascript
tricks. Most of our overrides to that file were also javascript tricks
as well so I'm moving them to where the rest of the donate form
javascript tricks are. Also updated them to work with changes to the
billable fields in this version of Civi.

Issue #23562