Cover some more billing selection scenarios
This still doesn't work completely right. The main problem I'm having is
with thing to trigger the billing check box events in CiviCRM. I think
instead of trying to reuse that code, I'm just going to have to copy it
because triggering it through events has too many side effects.