Fixed billing address copy when switiching processors
I tried a bunch of different things to trigger the CiviCRM code that
handles copying the shipping address to the billing address, but all of
that had annoying side effects. So I have just hard-coded in the field
names to copy. These will need to be updated if people mess with the
profiles, but I think that's unlikely.

Issue #23562