Fixed source field in Stripe case.
I think Stripe uses # as a kind of delimiter in the description and the
Strip plugin wants to put 'CiviCRM #' in the description sent to stripe to
maybe make it easier to find CiviCRM transactions. This causes the source
field in Civi to look a little funny. I don't think we need this on the
Civi side, so I changed it to only do that for the Strip description.

Issue #23562