Require first/last name fields for all contributions. Issue #24553.
There was an issue with PayPal transactions having blank name fields. This
was caused by a request to hide the name fields (along with the address
fields) for PayPal contributions when no premium was requested. To work
around this, we're now requiring the first/last name fields and have moved
them out of the Shiping Address fieldset.